Latest update

+++ src/main/java/org/bellatrix/services/LoadVAByEventResponse.java
@@ -9,7 +9,7 @@ public class LoadVAByEventResponse {
 
 	private List<VARecordView> vaRecord;
 	private ResponseStatus status;
-	private Long totalRecords;
+	private Integer totalRecords;
 
 	public List<VARecordView> getVaRecord() {
 		return vaRecord;
@@ -27,11 +27,11 @@ public class LoadVAByEventResponse {
 		this.status = status;
 	}
 
-	public Long getTotalRecords() {
+	public Integer getTotalRecords() {
 		return totalRecords;
 	}
 
-	public void setTotalRecords(Long totalRecords) {
+	public void setTotalRecords(Integer totalRecords) {
 		this.totalRecords = totalRecords;
 	}
 }
